fixed some issues with memberUpdate and connection

+++ events/messageCreate.js
@@ -59,11 +59,15 @@ module.exports = async (client, message) => {
         }
     }
 
-    //if client is mentioned, but no content is given return info
-    if (message.content.startsWith('<@') && message.content.endsWith('>')) {
-        if (message.mentions.users.first().id === client.user.id && messageArgs.length < 1) {
+
+    //check if message starts with
+    if (message.content.startsWith('<@')
+        && message.content.endsWith('>')
+        && message.member.permissions.has("KICK_MEMBERS")) {
+        const users = message.content.match(/[a-z\d]+/ig);
+        if (users[0] === client.user.id) {
             //reply with server info
-            message.reply(`Hello, your current server prefix is \`${prefix}\``)
+            message.reply(`Hello :wave:   Your current server prefix is \`${prefix}\``)
                 .then(msg => { setTimeout(() => msg.delete().catch((err) => { }), 4800) })
                 .catch((err) => { });
         }
